Extract persistent_device_id into DeviceResolver module
Previously lived on StorageVolume. Moving it to a shared module so
the encrypted swap setup script can reuse it.
